+++ Includes/LknmpMercadoPagoForGiveWPi18n.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Lkn\LknMercadoPagoForGiveWp\Includes;
+namespace Lknmp\MercadoPagoForGiveWp\Includes;
 /**
  * Define the internationalization functionality
  *
  * Loads and defines the internationalization files for this plugin
  * so that it is ready for translation.
  *
- * @link       https://https://www.linknacional.com.br/wordpress/givewp/
+ * @link       https://www.linknacional.com.br/wordpress/givewp/
  * @since      1.0.0
  *
- * @package    Lkn_Mercadopago_For_Givewp
- * @subpackage Lkn_Mercadopago_For_Givewp/includes
+ * @package    Lknmp_Mercadopago_For_Givewp
+ * @subpackage Lknmp_Mercadopago_For_Givewp/includes
  */
 
 /**
@@ -21,11 +21,11 @@ namespace Lkn\LknMercadoPagoForGiveWp\Includes;
  * so that it is ready for translation.
  *
  * @since      1.0.0
- * @package    Lkn_Mercadopago_For_Givewp
- * @subpackage Lkn_Mercadopago_For_Givewp/includes
+ * @package    Lknmp_Mercadopago_For_Givewp
+ * @subpackage Lknmp_Mercadopago_For_Givewp/includes
  * @author     Link Nacional <contato@linknacional>
  */
-final class LknMercadoPagoForGiveWPi18n {
+final class LknmpMercadoPagoForGiveWPi18n {
     /**
      * Load the plugin text domain for translation.
      *
@@ -33,7 +33,7 @@ final class LknMercadoPagoForGiveWPi18n {
      */
     public function load_plugin_textdomain(): void {
         load_plugin_textdomain(
-            'lkn-mercadopago-for-givewp',
+            'lknmp-mercadopago-for-givewp',
             false,
             dirname(dirname(plugin_basename(__FILE__))) . '/languages/'
         );