+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# 1.0.1 - 05/08/2024
+* Mensagem adicionada ao inserir o token do Mercado Pago;
+* Namespace do plugin atualizado;
+* Refatoração com as novas diretrizes do WordPress;
+* Atualização do README.
+
 # 1.0.0 - 31/05/2024
-* Plugin Launch.
+* Lançamento do plugin.